Refactor: 프로젝트 폴더 및 파일 구조 재정리
