+++ src/components/layout/PageLayout.tsx
@@ -1,4 +1,4 @@
-import Header from "@/components/gnb/CustomHeader";
+import Header from "@/components/Header/Header";
 import Footer from "@/components/Footer";
 import styled from "styled-components";
 import { useEffect, useState } from "react";
