+++ src/components/notice/noticeInfoCard.tsx
