--- src/data/mockPosts.ts
@@ -1,94 +1,0 @@
-import { PostProps } from "@/components/Postcard";  
-
-export const mockPosts: PostProps[] = [
-  {
-    id: 1,
-    imageUrl: "images/Postcard/restaurant1.svg",
-    name: "물방개맛집식당",
-    startsAt: "2025-01-02T15:00:00.000Z",
-    workhour: 4,
-    address1: "서울시 강남구",
-    hourlyPay:15200,
-    originalHourlyPay: 10320,
-  },
-  {
-    id: 2,
-    imageUrl: "images/Postcard/restaurant2.svg",
-    name: "맛집북까페",
-    startsAt: "2025-11-02T15:00:00.000Z",
-    workhour: 4,
-    address1: "서울시 강서구",
-    hourlyPay:22000,
-    originalHourlyPay: 10320,
-  },
-  {
-    id: 3,
-    imageUrl: "images/Postcard/restaurant3.svg",
-    name: "도토리식당",
-    startsAt: "2025-11-02T15:00:00.000Z",
-    workhour: 4,
-    address1: "서울시 강서구",
-    hourlyPay:15000,
-    originalHourlyPay: 10320,
-  },
-  {
-    id: 4,
-    imageUrl: "images/Postcard/restaurant4.svg",
-    name: "수리 에스프레소 맛집",
-    startsAt: "2025-11-02T15:00:00.000Z",
-    workhour: 4,
-    address1: "서울시 송파구",
-    hourlyPay:18000,
-    originalHourlyPay: 10320,
-  },
-  {
-    id: 5,
-    imageUrl: "images/Postcard/restaurant5.svg",
-    name: "별빛까페",
-    startsAt: "2025-12-02T15:00:00.000Z",
-    workhour: 4,
-    address1: "서울시 마포구",
-    hourlyPay:9900,
-    originalHourlyPay: 10320,
-  },
-  {
-    id: 6,
-    imageUrl: "images/Postcard/restaurant6.svg",
-    name: "커피바다",
-    startsAt: "2025-12-02T15:00:00.000Z",
-    workhour: 4,
-    address1: "서울시 광진구",
-    hourlyPay:11000,
-    originalHourlyPay: 10320,
-  },
-  {
-    id: 7,
-    imageUrl: "images/Postcard/restaurant7.svg",
-    name: "해피버거",
-    startsAt: "2025-12-02T15:00:00.000Z",
-    workhour: 4,
-    address1: "서울시 도봉구",
-    hourlyPay:9500,
-    originalHourlyPay: 10320,
-  },
-  {
-    id: 8,
-    imageUrl: "images/Postcard/restaurant8.svg",
-    name: "정원식당",
-    startsAt: "2025-12-02T15:00:00.000Z",
-    workhour: 4,
-    address1: "서울시 동대문구",
-    hourlyPay:10000,
-    originalHourlyPay: 10320,
-  },
-  {
-    id: 9,
-    imageUrl: "images/Postcard/restaurant9.svg",
-    name: "우리동네카페",
-    startsAt: "2025-11-02T15:00:00.000Z",
-    workhour: 4,
-    address1: "서울시 강서구",
-    hourlyPay:9500,
-    originalHourlyPay: 10320,
-  },
-];
