--- src/pages/alert.tsx
@@ -1,16 +1,0 @@
-import Head from "next/head";
-
-export default function Alert() {
-  return (
-    <>
-      <Head>
-        <title>알림 - theJulge</title>
-        <meta name="description" content="theJulge 알림 페이지" />
-      </Head>
-      <div>
-        <h1>알림</h1>
-        {/* 알림 리스트 컴포넌트가 들어갈 예정 */}
-      </div>
-    </>
-  );
-}
