--- src/pages/notice/edit.tsx
@@ -1,16 +1,0 @@
-import Head from "next/head";
-
-export default function EditNotice() {
-  return (
-    <>
-      <Head>
-        <title>공고 편집 - theJulge</title>
-        <meta name="description" content="공고 편집 페이지" />
-      </Head>
-      <div>
-        <h1>공고 편집</h1>
-        {/* 공고 편집 폼 컴포넌트가 들어갈 예정 */}
-      </div>
-    </>
-  );
-}
