--- src/pages/notice/register.tsx
@@ -1,16 +1,0 @@
-import Head from "next/head";
-
-export default function RegisterNotice() {
-  return (
-    <>
-      <Head>
-        <title>공고 등록 - theJulge</title>
-        <meta name="description" content="공고 등록 페이지" />
-      </Head>
-      <div>
-        <h1>공고 등록</h1>
-        {/* 공고 등록 폼 컴포넌트가 들어갈 예정 */}
-      </div>
-    </>
-  );
-}
