+++ src/pages/profile/edit.tsx
@@ -10,10 +10,9 @@ import {
   ButtonWrapper,
   SubmitButton,
   Label,
-} from "../../styles/profileEdit.styles";
+} from "./profileEdit.styles";
 
 import Modal from "../../components/Modal";
-import CustomHeader from "../../components/gnb/CustomHeader";
 import { updateMyProfile, fetchMyInfo } from "../../lib/api/user";
 
 export default function ProfileEditPage() {
@@ -75,7 +74,6 @@ export default function ProfileEditPage() {
 
   return (
     <PageWrapper>
-      <CustomHeader />
 
       <ProfileContainer>
         <Title>내 프로필</Title>
