--- src/pages/profile/register.tsx
@@ -1,16 +1,0 @@
-import Head from "next/head";
-
-export default function RegisterProfile() {
-  return (
-    <>
-      <Head>
-        <title>프로필 등록 - theJulge</title>
-        <meta name="description" content="알바 프로필 등록 페이지" />
-      </Head>
-      <div>
-        <h1>프로필 등록</h1>
-        {/* 프로필 등록 폼 컴포넌트가 들어갈 예정 */}
-      </div>
-    </>
-  );
-}
