--- src/styles/Home.module.css
@@ -1,10 +1,0 @@
-.main {
-  padding: 4rem 0;
-  text-align: center;
-}
-
-.title {
-  margin: 0;
-  line-height: 1.15;
-  font-size: 4rem;
-}
