--- src/styles/globalStyles.ts
@@ -1,44 +1,0 @@
-import { createGlobalStyle } from "styled-components";
-
-export const GlobalStyle = createGlobalStyle`
-  * {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-  }
-
-  html, body {
-    font-family: 'SpoqaHanSansNeo', sans-serif;
-    font-size: 16px;
-    color: #333;
-    background-color: #fff;
-    line-height: 1.5;
-  }
-
-  body.modal-open {
-    overflow: hidden;
-    -ms-overflow-style: none;  /* IE, Edge */
-    scrollbar-width: none;     /* Firefox */
-  }
-
-  body.modal-open::-webkit-scrollbar {
-    display: none;             /* Chrome, Safari, Opera */
-  }
-
-  a {
-    text-decoration: none;
-    color: inherit;
-  }
-
-  button {
-    cursor: pointer;
-    border: none;
-    background: none;
-  }
-
-  img {
-    display: block;
-    max-width: 100%;
-    height: auto;
-  }
-`;
