--- src/styles/globals.css
@@ -1,8 +1,0 @@
-html,
-body,
-* {
-  margin: 0;
-  padding: 0;
-  min-height: 1vh;
-  box-sizing: border-box;
-}
